Classim/requirements.txt
altgraph==0.16.1
asn1crypto==0.24.0
beautifulsoup4==4.7.1
certifi==2020.6.20
cffi==1.12.3
chardet==3.0.4
conda==4.9.0
conda-package-handling==1.3.10
cryptography==2.7
cycler==0.10.0
dateparser==0.7.4
future==0.17.1
idna==2.8
JSAnimation==0.1
kiwisolver==1.1.0
libarchive-c==2.8
macholib==1.11
matplotlib==3.1.3
menuinst==1.4.16
mkl-fft==1.0.12
mkl-random==1.0.2
mysql==0.0.2
mysqlclient==1.4.2.post1
mysql-connector-python==8.0.16
numpy==1.16.4
pandas==0.24.2
pefile==2019.4.18
pip==19.1.1
protobuf==3.6.0
psycopg2==2.8.6
ptvsd==4.2.10
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
PyInstaller==3.4
PyOpenGL==3.1.1a1
pyOpenSSL==19.0.0
pyparsing==2.4.0
PyQt5-sip==4.19.18
pyqtgraph==0.10.0
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
pywin32==223
pywin32-ctypes==0.2.0
regex==2020.2.20
requests==2.22.0
ruamel-yaml==0.15.46
scipy==1.3.0
setuptools==41.0.1
six==1.12.0
soupsieve==1.8
tornado==6.0.3
tqdm==4.32.1
tzlocal==2.0.0
urllib3==1.24.2
wheel==0.33.4
wincertstore==0.2
win-inet-pton==1.1.0
xmltodict==0.12.0
-yinstaller==3.4
